+++ src/components/trip-days-list.js
@@ -1,0 +1,18 @@
+const createTripDaysListTemplate = () => {
+  return (
+    `<ul class="trip-days"></ul>`
+  );
+};
+
+const createTripDayItemTemplate = () => {
+  return (
+    `<li class="trip-days__item  day">
+      <div class="day__info">
+        <span class="day__counter">1</span>
+        <time class="day__date" datetime="2019-03-18">MAR 18</time>
+      </div>
+    </li>`
+  );
+};
+
+export {createTripDaysListTemplate, createTripDayItemTemplate};
